honestly dont remember what i changed... fixed some bugs or something. I think I added some comments as well

+++ gate.h
@@ -5,14 +5,14 @@
 
 typedef struct Gate {
 	char *name;
-	int n_inputs;
-	int n_gates;
+	size_t n_inputs;
+	size_t n_gates;
 	char *operation; //in postfix
 } Gate;
 
 Gate *gate_parse(char *path);
 
-Gate *gate_new(const char* name, int n_inputs, const char *operation, int n_gates);
+Gate *gate_new(const char* name, int n_inputs, const char *operation, size_t n_gates);
 
 void gate_free(Gate *gate);
